updated saving server sample
