+++ SavingDashboards-Server/client/index.html
@@ -22,42 +22,40 @@
     <script src="https://cdn.jsdelivr.net/npm/jquery@3.6.0/dist/jquery.min.js"></script>
     <script src="https://cdn.quilljs.com/1.3.6/quill.min.js"></script>
     <script src="https://unpkg.com/dayjs@1.8.21/dayjs.min.js"></script>
-    <script src="https://dl.revealbi.io/reveal/libs/1.1.7/infragistics.reveal.js"></script>
+    <script src="https://dl.revealbi.io/reveal/libs/1.3.0/infragistics.reveal.js"></script>
 
     <script type="text/javascript">
         //set this to your server url
-        $.ig.RevealSdkSettings.setBaseUrl("https://localhost:7111/");
+        $.ig.RevealSdkSettings.setBaseUrl("http://localhost:5111/");
 
-        $.ig.RevealSdkSettings.ensureFontsLoadedAsync().then(() => {
-            $.ig.RVDashboard.loadDashboard("Sales2", (dashboard) => {
-                let revealView = new $.ig.RevealView("#revealView");
-                revealView.dashboard = dashboard;
+        $.ig.RVDashboard.loadDashboard("Sales").then(dashboard => {
+            let revealView = new $.ig.RevealView("#revealView");
+            revealView.dashboard = dashboard;
 
-                revealView.onSave = (rv, args) => {
-                    if (args.saveAs) {
-                        var newName = prompt("Please enter the dashboard name");
-                        isDuplicateName(newName).then(isDuplicate => {
-                            if (isDuplicate === "true") {
-                                if (!window.confirm("A dashboard with name: " + newName + " already exists. Do you want to override it?")) {
-                                    return;
-                                }
+            revealView.onSave = (rv, args) => {
+                if (args.saveAs) {
+                    var newName = prompt("Please enter the dashboard name");
+                    isDuplicateName(newName).then(isDuplicate => {
+                        if (isDuplicate === "true") {
+                            if (!window.confirm("A dashboard with name: " + newName + " already exists. Do you want to override it?")) {
+                                return;
                             }
+                        }
 
-                            args.dashboardId = args.name = newName;
-                            args.saveFinished();
-                        });
-
-                    }
-                    else {
+                        args.dashboardId = args.name = newName;
                         args.saveFinished();
-                    }
+                    });
+
                 }
-            });
+                else {
+                    args.saveFinished();
+                }
+            }
         });
 
 
         function isDuplicateName(name) {
-            return fetch(`https://localhost:7111/isduplicatename/${name}`).then(resp => resp.text());
+            return fetch(`http://localhost:5111/isduplicatename/${name}`).then(resp => resp.text());
         }
 
     </script>
