--- SavingDashboards-Server/server/RevealSdk.Server.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.0.31919.166
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RevealSdk.Server", "RevealSdk.Server\RevealSdk.Server.csproj", "{44921237-B370-420E-B6A6-936462439499}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{44921237-B370-420E-B6A6-936462439499}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{44921237-B370-420E-B6A6-936462439499}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{44921237-B370-420E-B6A6-936462439499}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{44921237-B370-420E-B6A6-936462439499}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {0B5AE6D5-9AED-4303-A8EB-3218BE221915}
-	EndGlobalSection
-EndGlobal
